backend/requirements.txt
fastapi
uvicorn
pandas
openai
pyyaml
python-multipart
python-dotenv
